requirements.txt
txtflask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
gunicorn==21.2.0
gevent==23.9.1